% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{\link[renv:init]{renv::init} with extra steps to maintain vscode-R features}
\usage{
init(
  project = NULL,
  ...,
  profile = NULL,
  settings = NULL,
  bare = FALSE,
  force = FALSE,
  restart = interactive()
)
}
\arguments{
\item{project}{The project directory. The \R working directory will be
changed to match the requested project directory.}

\item{...}{Unused arguments, reserved for future expansion. If any arguments
are matched to \code{...}, \code{renv} will signal an error.}

\item{profile}{The profile to be activated. When \code{NULL}, the default
profile is activated instead. See \code{vignette("profiles", package = "renv")}
for more information.}

\item{settings}{A list of \link[renv]{settings} to be used with the newly-initialized
project.}

\item{bare}{Boolean; initialize the project without attempting to discover
and install R package dependencies?}

\item{force}{Boolean; force initialization? By default, \code{renv} will refuse
to initialize the home directory as a project, to defend against accidental
mis-usages of \code{init()}.}

\item{restart}{Boolean; attempt to restart the \R session after initializing
the project? A session restart will be attempted if the \code{"restart"} \R
option is set by the frontend embedding \R.}
}
\description{
This function is a wrapper of \link[renv:init]{renv::init} which adds some extra steps
to maintain vscode-R's sessionWatcher and 'R: Launch RStudio Addin'
features after an \link{renv} environment gets activated.

Step 1: Create \code{dependencies.R} at the path that is determined by
\code{\link[=getwd]{getwd()}}, if not specifically provided, and add \verb{library(languageserver); library(jsonlite); library(rlang)} to the file.

Step2: Add \code{library(rstudioapi)} to \code{dependencies.R}, and then add
\code{options(vsc.rstudioapi = TRUE)} to \code{.Rprofile} created by \link[renv:init]{renv::init}.
}
\examples{

if (FALSE) init()
}
