% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvsc-package.R
\docType{package}
\name{renvsc-package}
\alias{renvsc}
\alias{renvsc-package}
\title{renvsc: [renv::init] with extra steps to maintain vscode-R features}
\description{
This function is a wrapper of [renv::init] which adds 
 some extra steps to maintain vscode-R's sessionWatcher and 
 'R: Launch RStudio Addin' features after an [renv] environment gets
 activated.
}
\author{
\strong{Maintainer}: Amarin Siripanich

}
\keyword{internal}
